restructured folder for contexts and hooks
