+++ src/app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from 'next'
 import '@/styles/globals.css'
 import { Manrope, Rubik } from 'next/font/google'
 import { Toaster } from 'react-hot-toast'
-import NextAuthSessionProvider from '@/shared/components/layout/NextAuthSessionProvider'
-import UserProvider from '@/shared/components/layout/UserProvider'
+import NextAuthSessionProvider from '@/shared/contexts/NextAuthSessionProvider'
+import UserProvider from '@/shared/contexts/UserProvider'
 
 import 'mapbox-gl/dist/mapbox-gl.css'
