+++ src/shared/contexts/UserProvider.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import React, { createContext, useContext, useEffect, useState } from 'react'
+import React, { createContext, useEffect, useState } from 'react'
 import { useSession } from 'next-auth/react'
 import { getCurrentUser } from '@/shared/services/user'
 import { User } from '@/shared/types/userTypes'
@@ -10,15 +10,7 @@ interface UserContextType {
   setUserData: React.Dispatch<React.SetStateAction<User | null>>
 }
 
-const UserContext = createContext<UserContextType | undefined>(undefined)
-
-export const useUser = () => {
-  const context = useContext(UserContext)
-  if (!context) {
-    throw new Error('useUser must be used within a UserProvider')
-  }
-  return context
-}
+export const UserContext = createContext<UserContextType | undefined>(undefined)
 
 const USER_DATA_KEY = 'userData'
